Reorganized repo + implementing mkdocs
