+++ .old/SOAPClient/HelloService.wsdl
@@ -1,0 +1,45 @@
+<definitions name = "HelloService"
+             targetNamespace = "http://localhost:8000/helloService?wsdl">
+
+    <message name = "SayHelloRequest">
+        <part name = "firstName" type = "xsd:string"/>
+    </message>
+
+    <message name = "SayHelloResponse">
+        <part name = "greeting" type = "xsd:string"/>
+    </message>
+
+    <portType name = "HelloPortType">
+        <operation name = "sayHello">
+            <input message = "tns:SayHelloRequest"/>
+            <output message = "tns:SayHelloResponse"/>
+        </operation>
+    </portType>
+
+    <binding name = "Hello_Binding" type = "tns:HelloPortType">
+        <soap:binding style = "rpc"
+                      transport = "http://schemas.xmlsoap.org/soap/http"/>
+        <operation name = "sayHello">
+            <soap:operation soapAction = "sayHello"/>
+            <input>
+                <soap:body
+                        encodingStyle = "http://schemas.xmlsoap.org/soap/encoding/"
+                        namespace = "urn:examples:helloservice"/>
+            </input>
+
+            <output>
+                <soap:body
+                        encodingStyle = "http://schemas.xmlsoap.org/soap/encoding/"
+                        namespace = "urn:examples:helloservice"/>
+            </output>
+        </operation>
+    </binding>
+
+    <service name = "HelloService">
+        <documentation>WSDL File for HelloService</documentation>
+        <port binding = "tns:Hello_Binding" name = "HelloPort">
+            <soap:address
+                    location = "http://localhost:8000/helloService?wsdl" />
+        </port>
+    </service>
+</definitions>
