+++ .travis.yml
@@ -1,7 +1,3 @@
 language: node_js
 node_js:
   - lts/*
-
-before_install:
-  - cd projetJS
-
