+++ controllers/index.js
@@ -1,0 +1,9 @@
+const user = require('./user');
+const calendar = require('./calendar');
+const event = require('./event');
+
+module.exports = {
+    calendar,
+    user,
+    event,
+};
