+++ package.json
@@ -1,9 +1,9 @@
 {
-  "name": "projetjs",
+  "name": "Calendar-API",
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "node ./bin/www",
+    "start": "node bin/www",
     "test": "mocha"
   },
   "dependencies": {
