--- projetJS/.sequelizerc
@@ -1,8 +1,0 @@
-const path = require('path');
-
-module.exports = {
-  "config": path.resolve('./config', 'config.json'),
-  "models-path": path.resolve('./models'),
-  "seeders-path": path.resolve('./seeders'),
-  "migrations-path": path.resolve('./migrations')
-};
