--- projetJS/controllers/index.js
@@ -1,9 +1,0 @@
-const user = require('./user');
-const calendar = require('./calendar');
-const event = require('./event');
-
-module.exports = {
-    calendar,
-    user,
-    event,
-};
